+++ ui/src/main/java/io/snabble/sdk/ui/SnabbleUI.java
@@ -1,12 +1,14 @@
 package io.snabble.sdk.ui;
 
 
+import androidx.appcompat.app.ActionBar;
 import io.snabble.sdk.Project;
 import io.snabble.sdk.utils.Logger;
 
 public class SnabbleUI {
     private static Project project;
     private static SnabbleUICallback uiCallback;
+    private static ActionBar actionBar;
 
     /**
      * Registers a globally used project for use with views.
@@ -20,7 +22,7 @@ public class SnabbleUI {
      * Registers a callback that the application needs to implement and respond accordingly.
      * <p>
      * Remember to unregister the callback when not used anymore, for example if you implement
-     * this callback in an Activity onCreate, remember to call unregisterUiCallbacks in onDestroy()
+     * this callback in an Activity onStart, remember to call unregisterUiCallbacks in onStop()
      * <p>
      * Calling this multiple times overwrites the old callback
      */
@@ -43,6 +45,31 @@ public class SnabbleUI {
         }
 
         return uiCallback;
+    }
+
+    /**
+     * Registers a ActionBar for suggested changes on the action bar.
+     * <p>
+     * Remember to unregister the ActionBar when not used anymore, for example if you set
+     * this ActionBar in an Activity onCreate, remember to call unregisterActionBarListener in onDestroy()
+     * <p>
+     * Calling this multiple times overwrites the old callback
+     */
+    public static void registerActionBar(ActionBar bar) {
+        actionBar = bar;
+    }
+
+    /**
+     * Unregisters a callback that was previously registered.
+     */
+    public static void unregisterActionBar(ActionBar bar) {
+        if (actionBar == bar) {
+            actionBar = null;
+        }
+    }
+
+    public static ActionBar getActionBar() {
+        return actionBar;
     }
 
     public static Project getProject() {