+++ ui/src/main/java/io/snabble/sdk/ui/checkout/CheckoutAbortedView.java
@@ -28,6 +28,10 @@ class CheckoutAbortedView extends FrameLayout {
     private void inflateView() {
         inflate(getContext(), R.layout.view_checkout_aborted, this);
 
+        if (SnabbleUI.getActionBar() != null) {
+            SnabbleUI.getActionBar().setTitle(R.string.Snabble_Checkout_error);
+        }
+
         findViewById(R.id.back).setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View v) {