+++ ui/src/main/java/io/snabble/sdk/ui/checkout/CheckoutView.java
@@ -35,7 +35,6 @@ public class CheckoutView extends FrameLayout implements Checkout.OnCheckoutStat
 
     private Checkout.State previousState;
     private DelayedProgressDialog progressDialog;
-    private boolean showPricetoPayInPaymentSelection = true;
 
     public CheckoutView(Context context) {
         super(context);
@@ -89,7 +88,6 @@ public class CheckoutView extends FrameLayout implements Checkout.OnCheckoutStat
         switch (state) {
             case REQUEST_PAYMENT_METHOD:
                 PaymentMethodView paymentMethodView = new PaymentMethodView(getContext());
-                paymentMethodView.setShowPriceToPay(showPricetoPayInPaymentSelection);
                 displayView(paymentMethodView);
                 break;
             case WAIT_FOR_APPROVAL:
@@ -203,10 +201,6 @@ public class CheckoutView extends FrameLayout implements Checkout.OnCheckoutStat
         application.unregisterActivityLifecycleCallbacks(activityLifecycleCallbacks);
 
         unregisterListeners();
-    }
-
-    public void setShowPriceToPayInPaymentSelection(boolean show) {
-        showPricetoPayInPaymentSelection = show;
     }
 
     private Application.ActivityLifecycleCallbacks activityLifecycleCallbacks =